refactor(observe): consolidate 30 flat files into 10, grouped by telemetry domain
The remaining files could not become subpackages. Every one of observe's twenty
mode handlers takes Deps and calls BuildResponseMetadata, clampLimit, a hint and a
summary builder. Splitting them by domain would mean exporting that whole kernel —
clampLimit, ResponseMetadata construction, ~8 summary builders, the log-entry
normalizers — as cross-package API, and then adding a facade in observe anyway,
because cmd/browser-agent (off-limits here) imports observe.GetBrowserErrors and
friends by name. That is the internal/queries trade: a bigger exported surface
bought with nothing but a file count. So the split is by file, not by package.

The old grouping was by *kind of thing* rather than by *what it serves*, and the
call graph says so:

- The four summary_builders_*.go files held nine build*Summary functions. Every
  single one had exactly one production caller, and never in the same file.
  buildHistorySummary lived in summary_builders_network_events.go while its only
  caller and its parameter type (historyEntry) were both in analysis_history.go.
  Each builder now sits with its handler.

- The analysis_* prefix meant nothing. analysis.go held network_waterfall,
  websocket_status, vitals and tabs; analysis_a11y.go, analysis_errors.go and
  analysis_history.go held three unrelated analyze-tool modes; analysis_screenshot.go
  held an observe mode. Nothing in the family referenced anything else in it.

- summarized_logs_{types,fingerprint,grouping}.go existed only to serve
  GetSummarizedLogs, and grouping.go is the only reader of LogGroup's unexported
  timestamps/sourceCounts fields — a split that could never have become a package
  boundary without exporting them.

Filename provenance was not independent evidence: git log --diff-filter=A shows
the handlers_*, analysis_* and summarized_logs_* prefixes were all minted by one
commit (58264122), and the summary_builders_* prefixes by another (594105b6).

New layout, by telemetry domain:

  deps.go            Deps interface + MaxObserveLimit/clampLimit
  filtering.go       level ranks, substring match, JSON-path body filter
  metadata.go        ResponseMetadata and its paginated spellings
  logs.go            errors, logs, extension_logs, error_clusters + their summaries
  summarized_logs.go the fingerprint/group/periodicity algorithm and its handler
  network.go         network_waterfall, network_bodies, websocket_events/status
  session.go         actions, transients, pilot, history, vitals, tabs, performance
  correlation.go     error_bundles and timeline — the modes that join streams on time
  page_state.go      page, storage, indexeddb, screenshot, accessibility — the modes
                     that ask the live tab rather than read a buffer

Tests follow the same layout: analysis_test.go and summary_builders_test.go, which
each straddled four of the new files, were split along the same lines; storage_test.go,
analysis_save_test.go and handlers_transients_test.go were renamed onto their new
homes. All 141 test functions are preserved, none added, none removed.

Behavior preservation is mechanical: the multiset of non-comment, non-import code
lines is 3,233 lines before and after and diffs to zero once whitespace and the
declared renames from the hints/idbquery extractions are normalized. All 34
exported symbols survive.
