--- internal/tools/observe/handlers_extension_logs.go
@@ -1,61 +1,0 @@
-// Purpose: Observe handlers for extension debug log access.
-
-package observe
-
-import (
-	"encoding/json"
-	"time"
-
-	"github.com/brennhill/Kaboom-Browser-AI-Devtools-MCP/internal/buffers"
-	"github.com/brennhill/Kaboom-Browser-AI-Devtools-MCP/internal/capture"
-	"github.com/brennhill/Kaboom-Browser-AI-Devtools-MCP/internal/mcp"
-)
-
-func buildExtensionLogEntries(allLogs []capture.ExtensionLog, limit int, level string, minLevel string) []map[string]any {
-	matched := buffers.ReverseFilterLimit(allLogs, func(entry capture.ExtensionLog) bool {
-		if level != "" && entry.Level != level {
-			return false
-		}
-		if minLevel != "" && LogLevelRank(entry.Level) < LogLevelRank(minLevel) {
-			return false
-		}
-		return true
-	}, limit)
-
-	logs := make([]map[string]any, len(matched))
-	for i, entry := range matched {
-		logs[i] = map[string]any{
-			"level":     entry.Level,
-			"message":   entry.Message,
-			"source":    entry.Source,
-			"category":  entry.Category,
-			"data":      entry.Data,
-			"timestamp": entry.Timestamp,
-		}
-	}
-	return logs
-}
-
-// GetExtensionLogs returns internal extension debug logs.
-func GetExtensionLogs(deps Deps, req mcp.JSONRPCRequest, args json.RawMessage) mcp.JSONRPCResponse {
-	var params struct {
-		Limit int    `json:"limit"`
-		Level string `json:"level"`
-	}
-	mcp.LenientUnmarshal(args, &params)
-	params.Limit = clampLimit(params.Limit, 100)
-
-	allLogs := deps.GetCapture().GetExtensionLogs()
-	logs := buildExtensionLogEntries(allLogs, params.Limit, params.Level, "")
-
-	var newestTS time.Time
-	if len(allLogs) > 0 {
-		newestTS = allLogs[len(allLogs)-1].Timestamp
-	}
-
-	return mcp.Succeed(req, "Extension logs", map[string]any{
-		"logs":     logs,
-		"count":    len(logs),
-		"metadata": BuildResponseMetadata(deps.GetCapture(), newestTS),
-	})
-}
