+++ internal/tools/observe/page_state_screenshot_test.go
