+++ internal/tools/observe/page_state_storage_test.go
